Add score input type for creating scores
