+++ StarWar/Types.php
@@ -4,11 +4,12 @@
 	use StarWar\Type\CharacterType;
 	use StarWar\Type\MovieType;
 	use StarWar\Type\QuoteType;
-	use StarWar\Type\QuoteInputType;
+	use StarWar\Type\ScoreInputType;
 	use StarWar\Type\QueryType;
 	use StarWar\Type\ScoreType;
 	use StarWar\Type\UserType;
 	use StarWar\Type\MutationType;
+	use StarWar\Type\MovieEnum;
 	use GraphQL\Type\Definition\ListOfType;
 	use GraphQL\Type\Definition\NonNull;
 	use GraphQL\Type\Definition\Type;
@@ -30,8 +31,8 @@
 		private static $movie;
 		/** @var QuoteType */
 		private static $quote;
-		/** @var QuoteInputType */
-		private static $quoteInput;
+		/** @var ScoreInputType */
+		private static $scoreInput;
 		/** @var QueryType */
 		private static $query;
 		/** @var ScoreType */
@@ -63,17 +64,17 @@
 		}
 
 		/**
-		 * @return QuoteInputType
-		 */
-		public static function quoteInput() {
-			return self::$quoteInput ?: (self::$quoteInput = new QuoteInputType());
-		}
-
-		/**
 		 * @return QueryType
 		 */
 		public static function query() {
 			return self::$query ?: (self::$query = new QueryType());
+		}
+
+		/**
+		 * @return ScoreInputType
+		 */
+		public static function scoreInput() {
+			return self::$scoreInput ?: (self::$scoreInput = new ScoreInputType());
 		}
 
 		/**
